Rename: statisticBy로 시작하는 파일 이름 statsBy로 변경
