+++ src/components/myPage/index.js
@@ -8,7 +8,8 @@ export { default as SubscriptionProviders } from './SubscriptionProviders';
 export { default as CurrentUnsubscriptionInfo } from './CurrentUnsubscriptionInfo';
 
 export { default as Statistics } from './Statistics';
-export { default as StatisticByProvider } from './StatisticByProvider';
+export { default as StatsByProvider } from './StatsByProvider';
+export { default as StatsByMonthly } from './StatsByMonthly';
 
 export { default as CollectionCategoryButton } from './CollectionCategoryButton';
 export { default as Collections } from './Collections';
