+++ src/hooks/index.js
@@ -1,4 +1,4 @@
-export { default as useStatisticsByProvider } from './useStatisticsByProvider';
+export { default as useStatsByProvider } from './useStatsByProvider';
 export { default as useColorScheme } from './useColorScheme';
 export { default as useObserver } from './useObserver';
 export { default as usePagination } from './usePagination';
